# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M-Sklad Restaurant Inventory & ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [kassa]
      summary: List cash register shifts, scoped by venue location
      parameters:
        - name: locationId
          in: query
          schema:
            type: integer
        - name: registerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed]
      responses:
        "200":
          description: Shifts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftListItem"
    post:
      operationId: openShift
      tags: [kassa]
      summary: Open a new shift on a register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenShiftBody"
      responses:
        "201":
          description: Opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    get:
      operationId: getShift
      tags: [kassa]
      summary: Get a single shift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}/close:
    patch:
      operationId: closeShift
      tags: [kassa]
      summary: Close a shift with counted cash (Z-report)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseShiftBody"
      responses:
        "200":
          description: Closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
...
  /products:
    get:
      operationId: listProducts
      tags: [venue]
      summary: List sellable products (menu) of a house, grouped for the POS grid
      parameters:
        - name: houseId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [venue]
      summary: Add a product to a house's menu (location_admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [venue]
      summary: Update a product (location_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [venue]
      summary: Delete a product (location_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      parameters:
        - name: locationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaffMember
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffMemberBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    patch:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffMemberBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaffMember
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        registerId:
          type: integer
        cashierName:
          type: string
        status:
          type: string
          enum: [open, closed]
        openingCash:
          type: string
        closingCashCounted:
          type: ["string", "null"]
        expectedCash:
          type: ["string", "null"]
        cashDifference:
          type: ["string", "null"]
        totalSalesCash:
          type: string
        totalSalesCard:
          type: string
        totalReturns:
          type: string
        totalCollected:
          type: string
        totalDeposited:
          type: string
        notes:
          type: ["string", "null"]
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - registerId
        - cashierName
        - status
        - openingCash
        - totalSalesCash
        - totalSalesCard
        - totalReturns
        - totalCollected
        - totalDeposited
        - openedAt
    ShiftListItem:
      allOf:
        - $ref: "#/components/schemas/Shift"
        - type: object
          properties:
            registerName:
              type: string
            houseId:
              type: integer
            houseName:
              type: string
            locationId:
              type: integer
            locationName:
              type: ["string", "null"]
    OpenShiftBody:
      type: object
      properties:
        registerId:
          type: integer
        cashierName:
          type: string
        openingCash:
          type: number
      required:
        - registerId
        - cashierName
    CloseShiftBody:
      type: object
      properties:
        closingCashCounted:
          type: number
        notes:
          type: string
      required:
        - closingCashCounted
    Product:
      type: object
      properties:
        id:
          type: integer
        houseId:
          type: integer
        category:
          type: string
        name:
          type: string
        price:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - houseId
        - category
        - name
        - price
        - isActive
        - sortOrder
        - createdAt
    CreateProductBody:
      type: object
      properties:
        houseId:
          type: integer
        category:
          type: string
        name:
          type: string
        price:
          type: number
        sortOrder:
          type: integer
      required:
        - houseId
        - category
        - name
        - price
    UpdateProductBody:
      type: object
      properties:
        category:
          type: string
        name:
          type: string
        price:
          type: number
        isActive:
          type: boolean
        sortOrder:
          type: integer
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - isActive
        - createdAt
    CreateStaffMemberBody:
      type: object
      properties:
        name:
          type: string
        position:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        locationId:
          type: ["integer", "null"]
      required:
        - name
    UpdateStaffMemberBody:
      type: object
      properties:
        name:
          type: string
        position:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        locationId:
          type: ["integer", "null"]
...